requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

nltk
pdfkit
rake-nltk
beautifulsoup4
urllib3
googlesearch-python
chardet
azure-functions
